+++ R/taxa.R
@@ -1,11 +1,11 @@
 #' Get taxa
 #'
+#' @description
 #' Gets the (unique) scientific names and associated taxonomic information from
 #' the observations of a Camera Trap Data Package object.
-#'
-#' - Duplicates are identified based on `scientificName` and, if present,
-#' `taxonID`. The row with the highest number of non-missing taxonomic values is
-#' kept. In case of ties, the last row is chosen.
+#' Duplicate taxa (i.e. with the same `scientificName`) are removed, retaining
+#' the taxon with (first) a `taxonID` and (second) the most taxonomic
+#' information.
 #'
 #' @inheritParams print.camtrapdp
 #' @return A [tibble::tibble()] data frame with the taxonomic information,
@@ -17,28 +17,47 @@
 #' taxa(x)
 taxa <- function(x) {
   check_camtrapdp(x)
-  taxa <- observations(x) %>%
+  taxa <-
+    observations(x) %>%
     dplyr::filter(!is.na(.data$scientificName)) %>%
     dplyr::select("scientificName", dplyr::starts_with("taxon.")) %>%
     dplyr::distinct() %>%
     dplyr::rename_with(~ sub("^taxon.", "", .x)) %>%
     dplyr::arrange(scientificName)
 
-  # Identify duplicates with the least information, to remove
-  duplicates_to_remove <- taxa %>%
-    dplyr::mutate(count_non_na = rowSums(!is.na(.))) %>%
-    {if ("taxonID" %in% names(taxa)) dplyr::group_by(., scientificName, taxonID)
-      else dplyr::group_by(., scientificName)} %>%
-    dplyr::filter(dplyr::n() > 1) %>%
-    dplyr::arrange(count_non_na) %>%
-    dplyr::slice_head(n = 1) %>%
-    dplyr::ungroup() %>%
-    dplyr::select(-count_non_na)
+  # Remove duplicates without taxonID
+  if ("taxonID" %in% names(taxa)) {
+    duplicates_without_taxonid <-
+      taxa %>%
+      dplyr::group_by(.data$scientificName) %>%
+      dplyr::filter(dplyr::n() > 1) %>%
+      dplyr::filter(is.na(.data$taxonID))
+    taxa <- dplyr::anti_join(
+      taxa,
+      duplicates_without_taxonid,
+      by = names(duplicates_without_taxonid)
+    )
+  }
 
-  # Remove duplicates
-  unique_taxa <-
+  # Remove duplicates with the least information
+  duplicates_with_least_info <-
     taxa %>%
-    dplyr::anti_join(duplicates_to_remove, by = names(duplicates_to_remove))
+    dplyr::mutate(columns_with_info = rowSums(!is.na(.))) %>%
+    dplyr::group_by(.data$scientificName) %>%
+    dplyr::filter(dplyr::n() > 1) %>%
+    dplyr::arrange(dplyr::desc(columns_with_info)) %>%
+    dplyr::slice_tail(n = -1) %>% # Remove first row from group (with most info)
+    dplyr::ungroup() %>%
+    dplyr::select(-columns_with_info)
+  taxa <- dplyr::anti_join(
+    taxa,
+    duplicates_with_least_info,
+    by = names(duplicates_with_least_info)
+  )
 
-  return(unique_taxa)
+  # Drop any columns that are empty (e.g as result of dropping duplicates)
+  cols_to_keep <- colSums(is.na(taxa)) != nrow(taxa)
+  taxa <- taxa[, cols_to_keep, drop = FALSE]
+
+  return(taxa)
 }