+++ tests/testthat/test-taxa.R
@@ -28,54 +28,99 @@ test_that("taxa() returns the expected rows (unique taxa)", {
   )
 })
 
-test_that("taxa() removes duplicates and keeps the rows with most taxnomical
+test_that("taxa() removes duplicates and keeps the rows with most taxonomical
           information", {
   skip_if_offline()
   x <- example_dataset()
-  x_without_taxonID <- x
-  obs <- data.frame(
-    observationID = c(1, 2, 3, 4, 5),
+
+  # Duplicate sci names: keep first record with taxonID
+  observations(x) <- data.frame(
+    observationID = c(
+      "delete:no_taxonID",
+      "keep",
+      "delete:exact_duplicate",
+      "delete:not_first",
+      "delete:not_first",
+      "keep:diff_name",
+      "delete:not_first"
+    ),
     scientificName = c(
       "Anas platyrhynchos",
       "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
       "Anas strepera",
-      "Anas strepera",
-      "Homo sapiens"
+      "Anas strepera"
     ),
     taxon.taxonID = c(
-      "https://www.checklistbank.org/dataset/COL2023/taxon/DGP6",
-      "https://www.checklistbank.org/dataset/COL2023/taxon/DGP6",
-      "https://www.checklistbank.org/dataset/COL2023/taxon/DGPL",
-      "https://www.checklistbank.org/dataset/COL2023/taxon/DGPL",
-      "https://www.checklistbank.org/dataset/COL2023/taxon/6MB3T"
-    ),
-    taxon.taxonRank = c("species", "species", "species", "species", "species"),
-    taxon.vernacularNames.eng = c(NA, "mallard", "gadwall", "gadwall", "human"),
-    taxon.vernacularNames.nld = c(
-      "wilde eend", NA, "krakeend", "krakeend", "mens"
-    ),
-    taxon.family = c(NA, NA, NA, "Anatidae", NA)
+      NA_character_,
+      "DGP6",
+      "DGP6",
+      "DGP6_1",
+      "DGP6_2",
+      "DGPL",
+      "DGPL_1"
+    )
   )
-  observations(x) <- obs
-  observations(x_without_taxonID) <- obs %>% dplyr::select(-taxon.taxonID)
-
   expected_taxa <- dplyr::tibble(
-    scientificName = c("Anas platyrhynchos", "Anas strepera", "Homo sapiens"),
-    taxonID = c(
-    "https://www.checklistbank.org/dataset/COL2023/taxon/DGP6",
-    "https://www.checklistbank.org/dataset/COL2023/taxon/DGPL",
-    "https://www.checklistbank.org/dataset/COL2023/taxon/6MB3T"
-  ),
-  taxonRank = c("species", "species", "species"),
-  vernacularNames.eng = c("mallard", "gadwall", "human"),
-  vernacularNames.nld = c(NA, "krakeend", "mens"),
-  family = c(NA, "Anatidae", NA)
+    scientificName = c("Anas platyrhynchos", "Anas strepera"),
+    taxonID = c("DGP6", "DGPL")
   )
-
   expect_identical(taxa(x), expected_taxa)
-  expect_identical(
-    taxa(x_without_taxonID),
-    dplyr::select(expected_taxa, -taxonID)
+
+  # Duplicate with different amount of info: keep with most
+  observations(x) <- data.frame(
+    observationID = c(
+      "delete:1_columns",
+      "delete:4_columns",
+      "keep:5_columns"
+    ),
+    scientificName = c(
+      "Anas platyrhynchos",
+      "Anas platyrhynchos",
+      "Anas platyrhynchos"
+    ),
+    taxon.taxonRank = c(
+      NA_character_,
+      "species",
+      "species"
+    ),
+    taxon.class = c(
+      NA_character_,
+      "Animalia",
+      NA_character_
+    ),
+    taxon.family = c(
+      NA_character_,
+      NA_character_,
+      "Anatidae"
+    ),
+    taxon.vernacularNames.eng = c(
+      NA_character_,
+      "mallard",
+      "mallard"
+    ),
+    taxon.vernacularNames.fra = c(
+      NA_character_,
+      NA_character_,
+      NA_character_
+    ),
+    taxon.vernacularNames.nld = c(
+      NA_character_,
+      NA_character_,
+      "wilde eend"
+    )
   )
+  expected_taxa <- dplyr::tibble(
+    scientificName = c("Anas platyrhynchos"),
+    taxonRank = c("species"),
+    # class: removed because duplicate with less columns has it
+    family = c("Anatidae"),
+    vernacularNames.eng = c("mallard"),
+    # vernacularNames.fra removed because no data
+    vernacularNames.nld = c("wilde eend")
+  )
+  expect_identical(taxa(x), expected_taxa)
 })